Add taxi path type/edge/runway-association editing and diagram click-to-filter selection
Taxi Paths grid gains Type, RunwayNumber+RunwayDesignator, LeftEdge/RightEdge
(distinct from the existing lighted flags), and CenterLine/CenterLineLighted,
all newly pulled from SimConnect's TAXI_PATH facility data and confirmed
against the local MSFS 2024 SDK docs. TaxiPathSegment.Type is promoted from a
raw int to a real TaxiPathType enum (now covering the SDK's full 0-8 range)
since it's a user-facing picker now, the same treatment VasiType/
ApproachLightSystemType got earlier; RunwayDesignator/LeftEdge/RightEdge get
their own new enums for the same reason. All six fields also show up
automatically in the read-only Airport Data tab via its existing reflection-
based tree.

Reworks the Edit tab diagram's taxiway interaction per user feedback: a
click (or Ctrl+click) now only selects and filters the Taxi Paths grid down
to the clicked path(s) — MainViewModel.VisibleTaxiPathEdits, recomputed on
every diagram-driven selection change — instead of immediately opening the
batch-edit popover. A plain click on empty diagram space clears the
selection; the popover only opens via an explicit right-click
(OpenTaxiwayBatchEditCommand/TaxiwayContextMenuCommand).

Fixes a regression introduced by that rework: reassigning the grid's
ItemsSource made WPF's DataGrid raise its own (asynchronous, multi-event)
SelectionChanged as a side effect, which fed back into
SyncTaxiwaySelectionFromRows and immediately un-selected whatever the
diagram click had just selected — so the grid filtered correctly but
nothing ever visibly highlighted, multi-select never stuck, and right-click
had no selection to open a popover for. Fixed in MainWindow.xaml.cs by
tracking the grid's ItemsSource reference and debouncing for a short settle
window after it changes, found by bisecting against the last commit and
instrumenting the live app rather than by inspection alone.

+++ AirportSmith.Tests/Services/AirportDiagramProjectorTests.cs
@@ -412,7 +412,7 @@ public class AirportDiagramProjectorTests
             a.TaxiNames.Add(taxiName);
             a.TaxiPaths.Add(new TaxiPathSegment
             {
-                Type = (int)TaxiPathType.Taxi,
+                Type = TaxiPathType.Taxi,
                 TaxiNameId = taxiName.Id,
                 StartXMeters = 0,
                 StartZMeters = 0,
@@ -450,7 +450,7 @@ public class AirportDiagramProjectorTests
     {
         var airport = Airport(a => a.TaxiPaths.Add(new TaxiPathSegment
         {
-            Type = (int)TaxiPathType.Taxi,
+            Type = TaxiPathType.Taxi,
             TaxiNameId = null,
             StartXMeters = 0,
             StartZMeters = 0,
@@ -471,7 +471,7 @@ public class AirportDiagramProjectorTests
         // crash or a stale label.
         var airport = Airport(a => a.TaxiPaths.Add(new TaxiPathSegment
         {
-            Type = (int)TaxiPathType.Taxi,
+            Type = TaxiPathType.Taxi,
             TaxiNameId = Guid.NewGuid(),
             StartXMeters = 0,
             StartZMeters = 0,
@@ -489,7 +489,7 @@ public class AirportDiagramProjectorTests
     {
         var airport = Airport(a => a.TaxiPaths.Add(new TaxiPathSegment
         {
-            Type = (int)TaxiPathType.Runway,
+            Type = TaxiPathType.Runway,
             StartXMeters = 0,
             StartZMeters = 0,
             EndXMeters = 200,
@@ -506,7 +506,7 @@ public class AirportDiagramProjectorTests
     {
         var airport = Airport(a => a.TaxiPaths.Add(new TaxiPathSegment
         {
-            Type = (int)TaxiPathType.Taxi,
+            Type = TaxiPathType.Taxi,
             StartXMeters = null,
             StartZMeters = null,
             EndXMeters = 200,
@@ -528,8 +528,8 @@ public class AirportDiagramProjectorTests
         // tab's diagram click map a shape back to the right TaxiPathEditViewModel.
         var airport = Airport(a =>
         {
-            a.TaxiPaths.Add(new TaxiPathSegment { Type = (int)TaxiPathType.Runway, StartXMeters = 0, StartZMeters = 0, EndXMeters = 10, EndZMeters = 0 });
-            a.TaxiPaths.Add(new TaxiPathSegment { Type = (int)TaxiPathType.Taxi, StartXMeters = 20, StartZMeters = 0, EndXMeters = 30, EndZMeters = 0 });
+            a.TaxiPaths.Add(new TaxiPathSegment { Type = TaxiPathType.Runway, StartXMeters = 0, StartZMeters = 0, EndXMeters = 10, EndZMeters = 0 });
+            a.TaxiPaths.Add(new TaxiPathSegment { Type = TaxiPathType.Taxi, StartXMeters = 20, StartZMeters = 0, EndXMeters = 30, EndZMeters = 0 });
         });
 
         var diagram = AirportDiagramProjector.Project(airport);
